Close task window on ESC keypress.
This brings the behavior for Tasque in-line with other GNOME
applet-style apps like Tomboy or the Hamster applet in allowing them to
be quickly called, worked in, and then closed again.

The code for the handler is pretty much straight ripped out of Tomboy,
where I had nothing to do with writing it.

Signed-off-by: Matt Enright <awickedshimmy@gmail.com>