Initial implementation of tooltip for the tray icon (bug #532915)
So..... Here's my initial stab at implementing this.

When digging into this, I noticed a few things that I had to tweak:

1) Application.Backend had lots of logic that just manipulated the Application
*instance*, so I moved this logic (with my additions) into an
Application.SetBackend (IBackend backend) instance method. Cool and the gang?

2) Most of the logic for filtering out what items where in "Today", "Tomorrow",
etc was in TaskGroup, which was a *GUI* thing, not a data/model thing. In order
to reuse the code, I created new classes TaskGroupModel, and
CompleteTaskGroupModel, to encapsulate that logic, and make TaskGroup and
CompleteTaskGroup make use of those models. That ok? Other suggestions?

3) Ideally, the "today_model", "tomorrow_model", private fields in
Tasque.Application could also be shared with the TaskGroup instances in the
TaskMainWindow. I've not done this yet in this patch, so there's some extra
memory/CPU used here that could be avoided. Worth addressing in this patch?
Suggestions on a nice approach to doing that?

Any other feedback welcome as well, of course.